requirements.txt
#Kivy==1.10.1
numpy==1.16.2
opencv-python==4.1.0.25
pyrealsense2==2.20.0.721
cython==0.29.10
open3d-python ==0.7.0.0
#pclpy==0.11.0 #note: only works python 3.6 - Windows
#scikit-build==0.10.0

#Kivy Dependencies
    #docutils==0.14
    #Pygments==2.3.1
    #pypiwin32==223
    #kivy.deps.sdl2==0.1.18
    #kivy.deps.glew==0.1.10